+++ packages/ListBox/src/store/Provider.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import uuidv4 from "uuid/v4";
+import { v4 as uuidv4 } from "uuid";
 import reducer from "./reducer";
 import { getDataOptions } from "../components/Option/helpers/optionState";