+++ packages/Uploader/src/helpers.js
@@ -1,4 +1,4 @@
-import uuidv4 from "uuid/v4";
+import { v4 as uuidv4 } from "uuid";
 import superagent from "superagent";
 import * as types from "./types";