+++ packages/helpers/src/index.js
@@ -1,4 +1,5 @@
 export * from "./customPropTypes";
+export { default as callAll } from "./callAll";
 export { default as extractChildren } from "./extractChildren";
 export { default as extractChildrenProps } from "./extractChildrenProps";
 export { default as isDevEnv } from "./isDevEnv";