+++ packages/Calendar/CHANGELOG.md
@@ -13,3 +13,9 @@ All notable changes to this project will be documented in this file.
 ## [0.1.15] - 2020-09-25
 
 - Updated: Change nanoId to version 2.1.7 to avoid ie 11 errors [@tristanjasper](https://github.com/tristanjasper).
+
+## [1.0.7-alpha.0] - 2021-02-01
+
+### Changed
+
+- Replace `nanoid` by `uuid`.